requirements.txt
langgraph>=0.5.0
langgraph-checkpoint-sqlite>=1.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
langchain-graph-retriever>=0.1.0
langchain-tavily>=0.2.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
networkx>=3.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
typing-extensions>=4.8.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
faiss-cpu>=1.7.0
numpy>=1.24.0
scikit-learn>=1.3.0
plotly>=5.15.0
pandas>=2.0.0
pyvis>=0.3.2
graphrag>=0.3.0
flask>=3.0.3
flask-cors>=4.0.1